backend/pyproject.toml
[project]
name = "topnet-backend"
version = "0.1.0"
description = "TopNet - Natural-Language Cloud Network Topology Copilot Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.6.0",
    "networkx>=3.2.0",
    "python-dotenv>=1.0.0",
    "boto3>=1.34.0",
    "httpx>=0.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "httpx>=0.26.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
